Remove unnecessary casts and fix ger yaml precision.

+++ clients/include/blas2/testing_spr2_strided_batched.hpp
@@ -55,7 +55,7 @@ void testing_spr2_strided_batched_bad_arg(const Arguments& arg)
         int64_t           incx        = 1;
         int64_t           incy        = 1;
         int64_t           batch_count = 2;
-        int64_t           A_size      = int64_t(N) * (N + 1) / 2;
+        int64_t           A_size      = N * (N + 1) / 2;
         hipblasStride     stridex     = N * incx;
         hipblasStride     strideA     = A_size;
         hipblasStride     stridey     = N * incy;
