+++ clients/include/blas2/testing_tpmv_batched.hpp
@@ -56,7 +56,7 @@ void testing_tpmv_batched_bad_arg(const Arguments& arg)
         int64_t            N           = 100;
         int64_t            incx        = 1;
         int64_t            batch_count = 2;
-        int64_t            A_size      = int64_t(N) * (N + 1) / 2;
+        int64_t            A_size      = N * (N + 1) / 2;
 
         device_batch_vector<T> dA(A_size, 1, batch_count);
         device_batch_vector<T> dx(N, incx, batch_count);
